refactor(lib/cli): new file architecture for lib CLI files
