+++ lib/cli/commons.js
