+++ lib/cli/npq.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const yargs = require('yargs')
-const cliCommons = require('./cliCommons')
+const cliCommons = require('./commons')
 
 const argv = yargs
   .version()
